+++ .ci/localhost_ansible_tests.py
@@ -15,10 +15,6 @@ with ci_lib.Fold('unit_tests'):
     ci_lib.run('./run_tests -v')
 
 
-with ci_lib.Fold('job_setup'):
-    os.chmod(ci_lib.TESTS_SSH_PRIVATE_KEY_FILE, int('0600', 8))
-
-
 with ci_lib.Fold('machine_prep'):
     # generate a new ssh key for localhost ssh
     if not os.path.exists(os.path.expanduser("~/.ssh/id_rsa")):